add repopulate_tx function to internal API
