+++ libwallet/src/internal/tx.rs
@@ -14,11 +14,6 @@
 
 //! Transaction building functions
 
-use crate::grin_core::libtx::{
-	build,
-	proof::{ProofBuild, ProofBuilder},
-	tx_fee,
-};
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
 use std::io::Cursor;
 use uuid::Uuid;
@@ -258,35 +253,6 @@ where
 	error!("ADDED OUTPUT");
 
 	Ok(context)
-}
-
-/// Repopulates output in the slate's tranacstion
-/// with outputs from the stored context
-/// change outputs and tx log entry
-pub fn repopulate_tx<'a, T: ?Sized, C, K>(
-	wallet: &mut T,
-	keychain_mask: Option<&SecretKey>,
-	slate: &mut Slate,
-	context: &Context,
-) -> Result<(), Error>
-where
-	T: WalletBackend<'a, C, K>,
-	C: NodeClient + 'a,
-	K: Keychain + 'a,
-{
-	/*let keychain = wallet.keychain(keychain_mask)?;
-	let mut body = &slate.tx_or_err_mut()?.body;*/
-	//let mut parts = vec![];
-	/*for (id, _, value) in &context.get_inputs() {
-		if let Some(i) = input {
-			if i.is_coinbase {
-				parts.push(build::coinbase_input(value, coin.key_id.clone()));
-			} else {
-				parts.push(build::input(value, coin.key_id.clone()));
-			}
-		}
-	}*/
-	Ok(())
 }
 
 /// Complete a transaction
